SISTEMA-WEB/backend/requirements.txt
Django==5.2
djangorestframework==3.16.0
django-cors-headers==4.7.0
python-decouple==3.8
asgiref==3.8.1
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.1

